have one Dto-type, drop functions, move into proxyIdentifier-file
